requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
langchain-community
numpy
pypdf
python-multipart
google-cloud-texttospeech
google-auth
mutagen